+++ src/Fauna/herbivore_cohort.cpp
@@ -10,10 +10,11 @@
 
 using namespace Fauna;
 
-HerbivoreCohort::HerbivoreCohort(const int age_days,
-                                 const double body_condition, const Hft* hft,
-                                 const Sex sex, const double ind_per_km2)
-    : HerbivoreBase(age_days, body_condition, hft, sex),
+HerbivoreCohort::HerbivoreCohort(
+    const int age_days, const double body_condition, const Hft* hft,
+    const Sex sex, const double ind_per_km2,
+    const ForageEnergyContent& metabolizable_energy)
+    : HerbivoreBase(age_days, body_condition, hft, sex, metabolizable_energy),
       ind_per_km2(ind_per_km2) {
   if (ind_per_km2 < 0.0)
     throw std::invalid_argument(
@@ -21,10 +22,10 @@ HerbivoreCohort::HerbivoreCohort(const int age_days,
         "ind_per_km2 <0.0");
 }
 
-HerbivoreCohort::HerbivoreCohort(const Hft* hft, const Sex sex,
-                                 const double ind_per_km2)
-    : HerbivoreBase(hft, sex),  // parent establishment constructor
-      ind_per_km2(ind_per_km2) {
+HerbivoreCohort::HerbivoreCohort(
+    const Hft* hft, const Sex sex, const double ind_per_km2,
+    const ForageEnergyContent& metabolizable_energy)
+    : HerbivoreBase(hft, sex, metabolizable_energy), ind_per_km2(ind_per_km2) {
   if (ind_per_km2 < 0.0)
     throw std::invalid_argument(
         "Fauna::HerbivoreIndividual::HerbivoreIndividual() "